requirements.txt
aiofiles==0.3.2
arrow==0.10.0
asn1crypto==0.23.0
async-timeout==2.0.0
cffi==1.11.2
cryptography==2.1.3
docopt==0.6.2
httptools==0.0.9
idna==2.6
motor==1.1
pycparser==2.18
pymongo==3.5.1
PySocks==1.6.7
python-dateutil==2.6.1
sanic==0.6.0
sanic-mongo==1.7.1
six==1.11.0
ua-parser==0.7.3
ujson==1.35
user-agents==1.1.0
uvloop==0.8.1
websockets==4.0.1
